requirements.txt
torch
torchvision
torchmetrics
numpy
wandb
pandas
Pyarrow
pillow
tqdm
vit_pytorch
packaging
opencv-python
opencv-python-headless